refactor: split managers and delegate level handling
