+++ tests/embargo.test.js
@@ -3,7 +3,8 @@ import { CharacterFactory, ItemFactory } from '../src/factory.js';
 import { EventManager } from '../src/managers/eventManager.js';
 import { MetaAIManager, STRATEGY } from '../src/managers/ai-managers.js';
 import { PathfindingManager } from '../src/managers/pathfindingManager.js';
-import { ItemManager, EquipmentManager } from '../src/managers/managers.js';
+import { ItemManager } from '../src/managers/itemManager.js';
+import { EquipmentManager } from '../src/managers/equipmentManager.js';
 import { TagManager } from '../src/managers/tagManager.js';
 import { CombatCalculator } from '../src/combat.js';
 import { MeleeAI } from '../src/ai.js';
